med_sense_backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
passlib[bcrypt]
python-dotenv
email-validator